+++ changelog.md
@@ -4,6 +4,8 @@ Codion Change Log
 ## 0.18.54
 ### is.codion.common.db
 - ConnectionPoolStatistics cleanup and refactoring.
+### is.codion.swing.common.ui
+- FrameBuilder.alwaysOnTop() and focusableWindowState() added.
 ### is.codion.framework.domain
 - EntityDefinition.Builder.descriptionResource() added.
 - AttributeDefinition.Builder.descriptionResource() added.